[REFACTOR] 코드 정리
- 불필요한 Preview 제거
- enum class 분리
- import 정리

Fixes: TICO-31
