+++ Mobile/PomoroDo/app/src/main/java/com/tico/pomorodo/ui/todo/view/TodoProfiles.kt
@@ -10,17 +10,12 @@ import androidx.compose.foundation.layout.height
 import androidx.compose.foundation.rememberScrollState
 import androidx.compose.material3.VerticalDivider
 import androidx.compose.runtime.Composable
-import androidx.compose.runtime.getValue
-import androidx.compose.runtime.mutableIntStateOf
 import androidx.compose.runtime.remember
-import androidx.compose.runtime.saveable.rememberSaveable
-import androidx.compose.runtime.setValue
 import androidx.compose.ui.Modifier
 import androidx.compose.ui.res.stringResource
-import androidx.compose.ui.tooling.preview.Preview
 import androidx.compose.ui.unit.dp
 import com.tico.pomorodo.R
-import com.tico.pomorodo.data.local.datasource.DataSource
+import com.tico.pomorodo.data.local.entity.User
 import com.tico.pomorodo.ui.common.view.ProfileVertical
 import com.tico.pomorodo.ui.theme.PomoroDoTheme
 
@@ -62,22 +57,3 @@ fun TodoProfileItems(userList: List<User>, selectedIndex: Int, onClicked: (Int) 
         }
     }
 }
-
-@Preview(showBackground = true)
-@Composable
-fun TodoProfilesPreview() {
-    var selectedProfileIndex by rememberSaveable {
-        mutableIntStateOf(-1)
-    }
-    PomoroDoTheme() {
-        TodoProfileItems(
-            userList = DataSource.userList,
-            selectedIndex = selectedProfileIndex,
-            onClicked = { selectedProfileIndex = it })
-    }
-}
-
-data class User(
-    val name: String,
-    val profileUrl: String? = null
-)
