+++ Mobile/PomoroDo/app/src/main/java/com/tico/pomorodo/ui/todo/view/TodoTopBar.kt
@@ -1,12 +1,9 @@
 package com.tico.pomorodo.ui.todo.view
 
 import androidx.compose.foundation.layout.PaddingValues
-import androidx.compose.foundation.layout.padding
 import androidx.compose.material3.ExperimentalMaterial3Api
 import androidx.compose.material3.HorizontalDivider
 import androidx.compose.material3.IconButton
-import androidx.compose.material3.Scaffold
-import androidx.compose.material3.Surface
 import androidx.compose.material3.TopAppBar
 import androidx.compose.material3.TopAppBarColors
 import androidx.compose.runtime.Composable
@@ -14,9 +11,7 @@ import androidx.compose.runtime.getValue
 import androidx.compose.runtime.mutableStateOf
 import androidx.compose.runtime.remember
 import androidx.compose.runtime.setValue
-import androidx.compose.ui.Modifier
 import androidx.compose.ui.graphics.Color
-import androidx.compose.ui.tooling.preview.Preview
 import androidx.compose.ui.unit.dp
 import com.tico.pomorodo.R
 import com.tico.pomorodo.ui.common.view.CustomDropdownMenuItem
@@ -46,7 +41,7 @@ fun TopBar(
         actions = {
             IconButton(onClick = { showMenu = true }) {
                 SimpleIcon(
-                    size = 24,
+                    size = 28,
                     imageVector = PomoroDoTheme.iconPack[IC_MORE_INFO]!!,
                     contentDescriptionId = R.string.content_ic_more_info
                 )
@@ -96,21 +91,5 @@ fun TodoDropDownMenu(
             },
             contentPadding = PaddingValues(10.dp)
         )
-    }
-}
-
-@Preview
-@Composable
-fun TopBarPreview() {
-    PomoroDoTheme() {
-        Scaffold(
-            topBar = {
-                TopBar(onManageCategoryClicked = {}, onAddCategoryClicked = {})
-            }
-        ) {
-            Surface(modifier = Modifier.padding(it)) {
-
-            }
-        }
     }
 }
